Added TestA2DFunction class. Significantly improved readability of example and added example for TestA2DFunction construction. Significantly improved documentation.
